requirements.txt
# VAR-Encoder dependencies

# Core
torch>=2.0.0
torchvision>=0.15.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Logging
wandb>=0.15.0
tqdm>=4.65.0

# Image processing
Pillow>=9.0.0

# Utilities
numpy>=1.24.0
